Refactor error handling and API handlers
- Moved error handling utilities from `zap/lib/error-handling` to `zap/lib/api/errors` for better organization.
- Updated all service files to import errors from the new location.
- Refactored API route handlers to use the new `withAuthenticatedApiHandler` from `zap/lib/api/handlers`.
- Removed the old error handling utilities and handlers to streamline the codebase.
- Ensured consistent error handling across all API routes and services.
