+++ core/src/app/api/(protected)/user/notifications/unsubscribe/route.ts
@@ -1,6 +1,6 @@
 import "server-only";
 
-import { withAuthenticatedApiHandler } from "@/zap/lib/error-handling/handlers";
+import { withAuthenticatedApiHandler } from "@/zap/lib/api/handlers";
 import { unsubscribeUserService } from "@/zap/services/push-notifications/unsubscribe-user.service";
 
 export const DELETE = withAuthenticatedApiHandler(async () => {
