+++ core/src/zap/rpc/procedures/ai.rpc.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { authMiddleware, base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/error-handling/handlers";
+import { withRpcHandler } from "@/zap/lib/api/handlers";
 import {
   InputDeleteAPIKeySchema,
   InputGetAPIKeySchema,
