+++ core/src/zap/rpc/procedures/feedbacks.rpc.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { authMiddleware, base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/error-handling/handlers";
+import { withRpcHandler } from "@/zap/lib/api/handlers";
 import { FeedbackSchema } from "@/zap/schemas/feedback.schema";
 import { getAverageRatingService } from "@/zap/services/feedbacks/get-average-rating.service";
 import { getUserFeedbackService } from "@/zap/services/feedbacks/get-user-feedback.service";
