+++ core/src/zap/rpc/procedures/push-notifications.rpc.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { authMiddleware, base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/error-handling/handlers";
+import { withRpcHandler } from "@/zap/lib/api/handlers";
 import { InputSubscribeUserSchema } from "@/zap/schemas/push-notifications.schema";
 import { subscribeUserService } from "@/zap/services/push-notifications/subscribe-user.service";
 import { unsubscribeUserService } from "@/zap/services/push-notifications/unsubscribe-user.service";
