+++ core/src/zap/rpc/procedures/users.rpc.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/error-handling/handlers";
+import { withRpcHandler } from "@/zap/lib/api/handlers";
 import { InputGetUserIdFromMailSchema } from "@/zap/schemas/users.schema";
 import { getNumberOfUsersService } from "@/zap/services/users/get-number-of-users.service";
 import { getUserIdFromMailService } from "@/zap/services/users/get-user-id-from-mail.service";
