+++ core/src/zap/rpc/procedures/waitlist.rpc.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { base } from "@/rpc/middlewares";
-import { withRpcHandler } from "@/zap/lib/error-handling/handlers";
+import { withRpcHandler } from "@/zap/lib/api/handlers";
 import { WaitlistSchema } from "@/zap/schemas/waitlist.schema";
 import { getNumberOfPeopleInWaitlistService } from "@/zap/services/waitlist/get-number-of-people-in-waitlist.service";
 import { submitWaitlistEmailService } from "@/zap/services/waitlist/submit-email.service";
