+++ core/src/zap/services/ai/update-ai-settings.service.ts
@@ -5,9 +5,9 @@ import { and, eq } from "drizzle-orm";
 import { db } from "@/db";
 import { userAISettings } from "@/db/schema";
 import { getApiSettingsForUserAndProviderQuery } from "@/zap/db/queries/ai.query";
+import { NotFoundError } from "@/zap/lib/api/errors";
 import { encryptionKeyHex } from "@/zap/lib/crypto";
 import { encrypt } from "@/zap/lib/crypto/encrypt";
-import { NotFoundError } from "@/zap/lib/error-handling/errors";
 import type { AIProviderId, ModelName } from "@/zap/types/ai.types";
 
 interface UpdateAISettingsContext {
