+++ core/src/zap/services/mails/send-verification-mail.service.ts
@@ -2,7 +2,7 @@ import "server-only";
 
 import { ZAP_DEFAULT_SETTINGS } from "@/zap.config";
 import { VerificationMail } from "@/zap/components/mails/verification.mail";
-import { MailError } from "@/zap/lib/error-handling/errors";
+import { MailError } from "@/zap/lib/api/errors";
 import { resend } from "@/zap/lib/resend/server";
 
 const from = ZAP_DEFAULT_SETTINGS.MAIL.FROM;
